refactor(MachOTestingSupport): hoist generic-specialization test environment into shared module
The actor-backed indexer cache, MachOImage handle, and descriptor-lookup
helpers used to live inline inside GenericSpecializationTests as a private
Environment protocol. A second SwiftInterfaceTests suite now needs the
same wiring, so promote it to GenericSpecializationTestingEnvironment in
MachOTestingSupport (with package-level visibility) and have the existing
nested suites adopt it. Also wires MachOSwiftSection and SwiftInterface
into MachOTestingSupport's target deps so the protocol can reference the
relevant types directly.
